rename cli modules from .m[tj]s to .[tj]s
