+++ cli/cli-command.d.ts
@@ -1,6 +1,6 @@
 import type Environment from 'yeoman-environment';
 
-import type EnvironmentBuilder from './environment-builder.mjs';
+import type EnvironmentBuilder from './environment-builder.js';
 
 /**
  * @param args - positional arguments, a varidic argument is an array at last position.
