+++ cli/index.ts
@@ -16,6 +16,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { createProgram, buildCommands, buildJHipster, runJHipster, printJHipsterLogo, done } from './program.mjs';
-export { logger } from './utils.mjs';
-export { default as EnvironmentBuilder } from './environment-builder.mjs';
+export { buildCommands, buildJHipster, createProgram, done, printJHipsterLogo, runJHipster } from './program.ts';
+export { logger } from './utils.js';
+export { default as EnvironmentBuilder } from './environment-builder.js';
