+++ cli/jhipster-command.js
