+++ cli/logo.js
